Render a newly created food item
This commit applies changes that add a newly created food to the DOM by appending a row of HTML beneath the header row in the foods display table.

This commit also changes the signature of the method HTMLHelper.toHtml to accept two strings as input (foodName, foodCalorie), instead of accepting a single food object as input.