Implement deleting a food
This commit applies changes that delete a food from the server and remove it from the front-end markup. There is now an event handler listening for a click on each delete button. When the delete button is clicked, the associated food's id is passed into the method Food#delete(id), which sends an HTTP DELETE to the server. If the delete on the server is successful, that food's row is removed from the foods table in the markup.

This commit also changes the signature of the method HTMLHelper.toHtml to additionally accept an id parameter. This was necessary to keep track of which food to delete when the delete button is clicked. The full signature is now HTMLHelper.toHtml(foodName, foodCalorie, id), where foodName is a string, foodCalorie is a number, and id is a number.