Implement adding foods to meals in the browser
This commit adds a function which is invoked as a callback to a click on a 'Add to Meal' button (such as 'Add to Breakfast' or 'Add to Lunch'). This function, Meal.addFoodsTo(meal), takes a single string which can have the value 'breakfast', 'lunch', 'dinner', or 'snack'. When this function is called, all of the foods whose checkbox is selected are added as foods within the HTML table of the specified meal.

This commit also changes the signature of HTMLHelper.toHtml() to take an object with key-value pairs for foodName, foodCalorie, id, and checkbox (which is a boolean to indicate whether to include checkboxes).