Improve file directory structure
This commit adds two directories to allow for better organization of the codebase. There is now an eventHandlers directory which contains the files that apply event handlers to the DOM. There is also a config directory, which currently only contains a file with the root URL of the production server.