Refactor to get calories from the server
Previously, we were getting calories from the DOM and summing them to get the total calories for a meal. This was hard to work with because putting the data in the DOM is asynchronous, so getting data from the DOM was unpredictable. Now, the calorie data is coming directly from the server, so we don't have to rely on the DOM being populated in order to work with the data.