Alternate sorting between descending, ascending, and no order
This commit finishes the story "Sorting Foods table by calories". When the user first clicks on the Calories header, the food results are sorted by calories in ascending order. On the next click, the food results are sorted in descending order. On the next click, the food results are shown in the original unsorted order again.