Implement sorting on the Foods page
This commit implements sorting by calories on the Foods page (similarly to the recent progress on the Meals page).

When the user first clicks on the Calories header, the food results are sorted by calories in ascending order. On the next click, the food results are sorted in descending order. On the next click, the food results are shown in the original unsorted order again.